requirements.txt
https://github.com/New-dev0/Telethon/archive/Artifact.zip
python-decouple
redis